+++ assembler/source/conversion_functions.hpp
@@ -2,6 +2,6 @@
 #pragma once
 #include <cinttypes>
 
-bool hexStrToUint8(const char* hexStr, uint8_t& value);
-bool binStrToUint8(const char* binStr, uint8_t& value);
-bool octStrToUint8(const char* octStr, uint8_t& value);
+bool hexStrToUint8(const char* hexStr, uint16_t& value);
+bool binStrToUint8(const char* binStr, uint16_t& value);
+bool octStrToUint8(const char* octStr, uint16_t& value);
